Added user related handling
Added interface for user
added the user class itself implementing said interface
added a user controller and routing file just to try to learn about request handling. It's not overly complicated but I'm still semi confused and am mostly surviving off of examples online.

the object oriented side of all of this is really where I shine, as for the other stuff I'm still learning so bare with me.